eventmux: add helpers_test.go
Move test helpers there, where we can reuse them for other broker
implementations.
